requirements.txt
django
djangorestframework
pillow
django-modeltranslation